some bigfixes
